requirements.txt
PyDirectInput==1.0.4
PyScreeze==0.1.29
PyAutoGUI==0.9.54
pywin32==306
dearpygui
discord_webhook==1.3.0
Flask==3.0.0
GetPixelColor==0.2.3
keyboard==0.13.5
matplotlib==3.8.2
mss==9.0.1
opencv_python
Pillow==9.5.0
psutil==5.9.6
pynput==1.7.6
PyScreeze==0.1.29
pytesseract==0.3.10
Requests==2.31.0
transformers==4.35.2